Add files via upload
I was able to calculate 0.10 accuracy via linear regression model and i found that via feature selection if we select top 40 columns(see my code which cols) we will get the same accuracy as given by all the columns.
As accuracy via this model is too low. I don't think it is a good model to be considered for our analysis.